refactor: move org-membership-cache to auth layer
Move org-membership-cache.ts and its test from zero/org/ to auth/,
eliminating the auth→zero reverse dependency. Update all 4 consumer
import paths accordingly.

Closes #7790
