+++ turbo/apps/web/src/lib/auth/get-auth-context.ts
@@ -9,7 +9,7 @@ import {
   verifyZeroToken,
   verifyCliToken,
 } from "./sandbox-token";
-import { verifyMembershipCached } from "../zero/org/org-membership-cache";
+import { verifyMembershipCached } from "./org-membership-cache";
 import { logger } from "../logger";
 
 const log = logger("auth:user");
