+++ turbo/apps/web/src/lib/zero/org/resolve-org.ts
@@ -9,7 +9,7 @@ import {
 import { getOrgMetadata, getOrgData } from "./org-cache-service";
 import { orgMetadata } from "../../../db/schema/org-metadata";
 import { orgCache } from "../../../db/schema/org-cache";
-import { verifyMembershipCached } from "./org-membership-cache";
+import { verifyMembershipCached } from "../../auth/org-membership-cache";
 import type { AuthContext } from "../../auth/get-auth-context";
 
 import type { OrgRole } from "@vm0/core";
